replace with typescript implementation
