+++ .editorconfig
@@ -3,11 +3,5 @@ root = true
 [*]
 insert_final_newline = true
 trim_trailing_whitespace = true
-
-[*.{html,ejs,js,scss}]
-indent_style = tab
-indent_size = 4
-
-[*.{json,frag,vert}]
 indent_style = space
 indent_size = 2
