+++ .gitignore
@@ -1,5 +1,19 @@
+# OS
 .DS_Store
 
+# Editor
+.vscode
+
+# Cache
+.rpt2_cache
+
+# Node
 node_modules
+
+# Directories
+dist
 input
 output
+
+# Logging
+*.log
