+++ README.md
@@ -1,24 +1,20 @@
 # Binpacker
 
 [![npm version](https://badge.fury.io/js/%40timvanscherpenzeel%2Fbinpacker.svg)](https://www.npmjs.com/package/@timvanscherpenzeel/binpacker)
-[![dependencies](https://david-dm.org/timvanscherpenzeel/binpacker.svg)](https://david-dm.org/timvanscherpenzeel/binpacker)
-[![devDependencies](https://david-dm.org/timvanscherpenzeel/binpacker/dev-status.svg)](https://david-dm.org/timvanscherpenzeel/binpacker#info=devDependencies)
 
 CLI tool for packing multiple files into a single Binpack binary in order to save network requests in the browser.
 
 Inspired by [GLB File Format Specification](https://github.com/KhronosGroup/glTF/tree/master/specification/2.0#glb-file-format-specification), [MM.Loader](https://github.com/MM56/MM.Loader), [MM.Packer](https://github.com/MM56/mm-packer), [Magipack.js](https://github.com/keitakun/Magipack.js) and [this Twitter thread](https://twitter.com/tvscherpenzeel/status/1015124298812489728).
 
-## Live demo
-
-[Live demo](https://timvanscherpenzeel.github.io/binpacker-loader/) of [binpack-loader](https://github.com/timvanscherpenzeel/binpacker-loader) to efficiently extract packed files in a web worker from a single Binpack binary.
-
 ## Installation
+
+Make sure you have [Node.js](http://nodejs.org/) installed.
 
 ```sh
 $ npm install -g --save @timvanscherpenzeel/binpacker
 ```
 
-## Example
+## CLI Usage
 
 ```sh
 $ node ./bin/binpacker.js -i ./input -o ./output/example.binpack
@@ -85,9 +81,8 @@ The binary chunk (`ubyte[] chunkData`) contains a single `Uint8Array` typed arra
 ### Required
 
     -i, --input [example: ./input (glob) or ./input/manifest.json (manifest)] [required]
-
     -o, --output [example: ./output/example.binpack] [required]
 
 ## License
 
-`binpacker` is released under the [MIT license](https://raw.githubusercontent.com/TimvanScherpenzeel/binpacker/master/LICENSE).
+My work is released under the [MIT license](https://raw.githubusercontent.com/TimvanScherpenzeel/binpacker/master/LICENSE).
