+++ bin/binpacker.js
@@ -1,2 +1,3 @@
 #!/usr/bin/env node
-require('../lib/index.js');
+
+require('../dist/cli/lib/index.js').pack();
