--- lib/constants.js
@@ -1,32 +1,0 @@
-const constants = {
-  SUPPORTED_INPUT_TYPES: [
-    // Image formats
-    '.png',
-    '.jpg',
-    '.webp',
-    '.gif',
-
-    // Compressed image formats
-    '.ktx',
-    '.dds',
-    '.pvr',
-
-    // Audio
-    '.ogg',
-    '.mp3',
-
-    // Video
-    '.webm',
-    '.mp4',
-
-    // Data
-    '.css',
-    '.json',
-    '.txt',
-    '.glsl',
-    '.frag',
-    '.vert',
-  ],
-};
-
-module.exports = constants;
