+++ lib/constants.ts
@@ -1,0 +1,28 @@
+export const SUPPORTED_INPUT_TYPES = [
+  // Image formats
+  '.png',
+  '.jpg',
+  '.webp',
+  '.gif',
+
+  // Compressed image formats
+  '.ktx',
+  '.dds',
+  '.pvr',
+
+  // Audio
+  '.ogg',
+  '.mp3',
+
+  // Video
+  '.webm',
+  '.mp4',
+
+  // Data
+  '.css',
+  '.json',
+  '.txt',
+  '.glsl',
+  '.frag',
+  '.vert',
+];
