+++ tsconfig.json
@@ -1,0 +1,20 @@
+{
+  "compilerOptions": {
+    "moduleResolution": "node",
+    "target": "es2015",
+    "module": "commonjs",
+    "lib": ["es7"],
+    "strict": true,
+    "sourceMap": true,
+    "declaration": true,
+    "esModuleInterop": true,
+    "allowSyntheticDefaultImports": true,
+    "noImplicitAny": true,
+    "resolveJsonModule": true,
+    "declarationDir": "dist/types",
+    "outDir": "dist/cli",
+    "typeRoots": ["node_modules/@types"]
+  },
+  "include": ["lib/**/*.ts"],
+  "exclude": ["node_modules"]
+}
